fixed muted channel

+++ src/ags/audio/recall/ags_play_channel.c
@@ -219,7 +219,7 @@ ags_play_channel_init(AgsPlayChannel *play_channel)
 					     "control-port\0", ags_play_channel_control_port[0],
 					     "port-value-is-pointer\0", FALSE,
 					     "port-value-type\0", G_TYPE_UINT64,
-					     "port-value-size\0", sizeof(guint),
+					     "port-value-size\0", sizeof(guint64),
 					     "port-value-length\0", 1,
 					     NULL);
   play_channel->audio_channel->port_value.ags_port_uint = 0;
